Move override_num_splits_if_necessary into cpp
